Sort package inventory for versions script
The software-versions-inventory.sh script gets a nicely formatted list
of packages for each Bottlerocket version and renders it into a easily
readable markdown table with a lot of good details. This table currently
ends up listing the packages in whatever order the `find` command
returns the package names from intrspecting the repo contents, which
means it is a somewhat randomized list of packages that requires
searching through if you are looking for a specific item.

This is a minor change to sort the results of the `find` command so that
this table is always in alphabetical order, making it a little easier to
find what you're looking for.

Signed-off-by: Sean McGinnis <stmcg@amazon.com>